+++ Cargo.toml
@@ -10,4 +10,5 @@ members = [
 
 exclude = [
     "media/book/tests",
+    "argmin-math-ndarray-linalg-tests",
 ]
