+++ argmin-math-ndarray-linalg-tests/ndarray_0_13/src/lib.rs
@@ -1,0 +1,51 @@
+mod add {
+    include!(concat!(env!("CARGO_MANIFEST_DIR"), "/../src/add.rs"));
+}
+mod conj {
+    include!(concat!(env!("CARGO_MANIFEST_DIR"), "/../src/conj.rs"));
+}
+mod div {
+    include!(concat!(env!("CARGO_MANIFEST_DIR"), "/../src/div.rs"));
+}
+mod dot {
+    include!(concat!(env!("CARGO_MANIFEST_DIR"), "/../src/dot.rs"));
+}
+mod eye {
+    include!(concat!(env!("CARGO_MANIFEST_DIR"), "/../src/eye.rs"));
+}
+mod inv {
+    include!(concat!(env!("CARGO_MANIFEST_DIR"), "/../src/inv.rs"));
+}
+mod l1norm {
+    include!(concat!(env!("CARGO_MANIFEST_DIR"), "/../src/l1norm.rs"));
+}
+mod l2norm {
+    include!(concat!(env!("CARGO_MANIFEST_DIR"), "/../src/l2norm.rs"));
+}
+mod minmax {
+    include!(concat!(env!("CARGO_MANIFEST_DIR"), "/../src/minmax.rs"));
+}
+mod mul {
+    include!(concat!(env!("CARGO_MANIFEST_DIR"), "/../src/mul.rs"));
+}
+mod random {
+    include!(concat!(env!("CARGO_MANIFEST_DIR"), "/../src/random.rs"));
+}
+mod scaledadd {
+    include!(concat!(env!("CARGO_MANIFEST_DIR"), "/../src/scaledadd.rs"));
+}
+mod scaledsub {
+    include!(concat!(env!("CARGO_MANIFEST_DIR"), "/../src/scaledsub.rs"));
+}
+mod signum {
+    include!(concat!(env!("CARGO_MANIFEST_DIR"), "/../src/signum.rs"));
+}
+mod sub {
+    include!(concat!(env!("CARGO_MANIFEST_DIR"), "/../src/sub.rs"));
+}
+mod transpose {
+    include!(concat!(env!("CARGO_MANIFEST_DIR"), "/../src/transpose.rs"));
+}
+mod zero {
+    include!(concat!(env!("CARGO_MANIFEST_DIR"), "/../src/zero.rs"));
+}
