+++ argmin-math-ndarray-linalg-tests/ndarray_0_15/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "ndarray_0_15"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dev-dependencies]
+argmin-math = { path = "../../argmin-math", version = "0.3", features = [
+    "ndarray_v0_15",
+] }
+ndarray = { version = "0.15", default-features = false }
+ndarray-linalg = { version = "0.16", default-features = false, features = ["intel-mkl"] }
+num-complex = { version = "0.4", default-features = false }
+num-traits = { version = "0.2", default-features = false }
+num-integer = { version = "0.1", default-features = false }
+paste = "1"
+approx = "0.5.0"
+rand = "*" # Should unify with whatever is currently used in argmin itself
+
+[features]
+serde = ["argmin-math/ndarray_v0_15-serde"]
