+++ argmin-math/src/ndarray_m/l1norm.rs
@@ -67,78 +67,10 @@ make_l1norm_complex!(Complex<usize>, usize);
 make_l1norm_complex!(Complex<f32>, f32);
 make_l1norm_complex!(Complex<f64>, f64);
 
+// The tests expect the name for the crate containing the tested functions to be argmin_math
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use ndarray::{array, Array1};
-    use paste::item;
-
-    macro_rules! make_test {
-        ($t:ty) => {
-            item! {
-                #[test]
-                fn [<test_norm_ $t>]() {
-                    let a = array![4 as $t, 3 as $t];
-                    let res = <Array1<$t> as ArgminL1Norm<$t>>::l1_norm(&a);
-                    let target = 7 as $t;
-                    assert!(((target - res) as f64).abs() < std::f64::EPSILON);
-                }
-            }
-
-            item! {
-                #[test]
-                fn [<test_norm_complex_ $t>]() {
-                    let a = array![Complex::new(4 as $t, 2 as $t), Complex::new(3 as $t, 4 as $t)];
-                    let res = <Array1<Complex<$t>> as ArgminL1Norm<$t>>::l1_norm(&a);
-                    let target = a[0].l1_norm() + a[1].l1_norm();
-                    assert!(((target - res) as f64).abs() < std::f64::EPSILON);
-                }
-            }
-        };
-    }
-
-    macro_rules! make_test_signed {
-        ($t:ty) => {
-            item! {
-                #[test]
-                fn [<test_norm_signed_ $t>]() {
-                    let a = array![-4 as $t, -3 as $t];
-                    let res = <Array1<$t> as ArgminL1Norm<$t>>::l1_norm(&a);
-                    let target = 7 as $t;
-                    assert!(((target - res) as f64).abs() < std::f64::EPSILON);
-                }
-            }
-
-            item! {
-                #[test]
-                fn [<test_norm_signed_complex_ $t>]() {
-                    let a = array![Complex::new(-4 as $t, -2 as $t), Complex::new(-3 as $t, -4 as $t)];
-                    let res = <Array1<Complex<$t>> as ArgminL1Norm<$t>>::l1_norm(&a);
-                    let target = a[0].l1_norm() + a[1].l1_norm();
-                    assert!(((target - res) as f64).abs() < std::f64::EPSILON);
-                }
-            }
-        };
-    }
-
-    make_test!(isize);
-    make_test!(usize);
-    make_test!(i8);
-    make_test!(u8);
-    make_test!(i16);
-    make_test!(u16);
-    make_test!(i32);
-    make_test!(u32);
-    make_test!(i64);
-    make_test!(u64);
-    make_test!(f32);
-    make_test!(f64);
-
-    make_test_signed!(isize);
-    make_test_signed!(i8);
-    make_test_signed!(i16);
-    make_test_signed!(i32);
-    make_test_signed!(i64);
-    make_test_signed!(f32);
-    make_test_signed!(f64);
-}
+use crate as argmin_math;
+include!(concat!(
+    env!("CARGO_MANIFEST_DIR"),
+    "/../argmin-math-ndarray-linalg-tests/src/l1norm.rs"
+));
