+++ argmin-math/src/ndarray_m/transpose.rs
@@ -43,78 +43,10 @@ make_add!(f64);
 make_add!(Complex<f32>);
 make_add!(Complex<f64>);
 
+// The tests expect the name for the crate containing the tested functions to be argmin_math
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use ndarray::array;
-    use paste::item;
-
-    macro_rules! make_test {
-        ($t:ty) => {
-            item! {
-                // not sure if this is such a smart test :D
-                #[test]
-                fn [<test_transpose_ $t>]() {
-                    let a = array![1 as $t, 4 as $t];
-                    let target = array![1 as $t, 4 as $t];
-                    let res = <Array1<$t> as ArgminTranspose<Array1<$t>>>::t(a);
-                    for i in 0..2 {
-                        assert!(((target[i] - res[i]) as f64).abs() < std::f64::EPSILON);
-                    }
-                }
-            }
-
-            item! {
-                #[test]
-                fn [<test_transpose_2d_1_ $t>]() {
-                    let a = array![
-                        [1 as $t, 4 as $t],
-                        [8 as $t, 7 as $t]
-                    ];
-                    let target = array![
-                        [1 as $t, 8 as $t],
-                        [4 as $t, 7 as $t]
-                    ];
-                    let res = <Array2<$t> as ArgminTranspose<Array2<$t>>>::t(a);
-                    for i in 0..2 {
-                        for j in 0..2 {
-                            assert!(((target[(i, j)] - res[(i, j)]) as f64).abs() < std::f64::EPSILON);
-                        }
-                    }
-                }
-            }
-
-            item! {
-                #[test]
-                fn [<test_transpose_2d_2_ $t>]() {
-                    let a = array![
-                        [1 as $t, 4 as $t],
-                        [8 as $t, 7 as $t],
-                        [3 as $t, 6 as $t]
-                    ];
-                    let target = array![
-                        [1 as $t, 8 as $t, 3 as $t],
-                        [4 as $t, 7 as $t, 6 as $t]
-                    ];
-                    let res = <Array2<$t> as ArgminTranspose<Array2<$t>>>::t(a);
-                    for i in 0..2 {
-                        for j in 0..3 {
-                            assert!(((target[(i, j)] - res[(i, j)]) as f64).abs() < std::f64::EPSILON);
-                        }
-                    }
-                }
-            }
-        };
-    }
-
-    make_test!(i8);
-    make_test!(u8);
-    make_test!(i16);
-    make_test!(u16);
-    make_test!(i32);
-    make_test!(u32);
-    make_test!(i64);
-    make_test!(u64);
-    make_test!(f32);
-    make_test!(f64);
-}
+use crate as argmin_math;
+include!(concat!(
+    env!("CARGO_MANIFEST_DIR"),
+    "/../argmin-math-ndarray-linalg-tests/src/transpose.rs"
+));
